+++ Party-Inc/Assets/PartyInc/Resources/2_Games/1_Egg/_Scripts/_general/Mono_FinishScreenUI_EGG.cs
@@ -136,7 +136,7 @@ namespace PartyInc
             {
                 PhotonNetwork.LeaveRoom();
 
-                SceneManager.LoadScene(Stt_SceneIndexes.HUB);
+                SceneManager.LoadScene((int)Stt_SceneIndexes.HUB);
             }
         }
     }
