+++ Party-Inc/Assets/PartyInc/Resources/2_Games/3_TumultousTap/_Scripts/_general/Mono_FinishScreenUI_TT.cs
@@ -167,7 +167,7 @@ namespace PartyInc
             {
                 PhotonNetwork.LeaveRoom();
 
-                SceneManager.LoadScene(Stt_SceneIndexes.HUB);
+                SceneManager.LoadScene((int)Stt_SceneIndexes.HUB);
             }
         }
     }
