+++ Party-Inc/Assets/PartyInc/Resources/4_Hub/_Scripts/Mono_GameResults.cs
@@ -242,7 +242,7 @@ namespace PartyInc
             {
                 PhotonNetwork.LeaveRoom();
 
-                SceneManager.LoadScene(Stt_SceneIndexes.HUB);
+                SceneManager.LoadScene((int)Stt_SceneIndexes.HUB);
             }
         }
     }
